Release 1.0.1: recording session, notifications, and UI polish
- Bump version to 1.0.1 (versionCode 16) for Play Store.

- Recording: session repository layer, interruption notifier, state transitions; fix duplicate idle notification while saving (isSaving) and guard stop double-tap.

- UI: capture mode pill uses theme surface; main tabs show icon-only when inactive and expand label when selected; Support credits and Play promo via billing flow.

- Notifications: consolidated app controls layouts; launcher mipmaps path update; strings and tooling updates.

+++ app/src/main/java/com/ibbie/catrec_screenrecorcer/service/BrushOverlayLayout.kt
@@ -754,7 +754,7 @@ internal class BrushOverlayLayout(
                     setColor(c)
                     setStroke(
                         dp(if (selected) 3 else 1),
-                        if (selected) Color.WHITE else 0x66FFFFFF.toInt(),
+                        if (selected) Color.WHITE else 0x66FFFFFF,
                     )
                 }
         }
